requirements.txt
pandas>=1.3.0
pandasai>=1.4.0
openai>=1.0.0
python-dotenv>=1.0.1
streamlit==1.24.0
Pillow